new charts

+++ js/chart.js
@@ -791,10 +791,7 @@ function updateTrendCharts() {
     // Format month labels
     const formatMonth = (monthStr) => {
         const date = new Date(monthStr + '-01');
-        return date.toLocaleDateString('en-US', { 
-            month: 'long',
-            year: 'numeric'
-        });
+        return date.toLocaleDateString('en-US', { month: 'short', year: 'numeric' });
     };
 
     // Seasonal Trends Chart
